Move required_region_bounds to rustc::infer::opaque_types.
